Remove a stale dependency.
It was introduced in https://github.com/SpineEventEngine/base/commit/6589305b5f84f81cfa8328d3dc61e181a76ea5e3
to make sure that the proto messages declared in `main` of `base/base`
are visible in the tests of the proto-js-plugin.

Removing this dependency does not cause the messages to become invisible.